Allow CardMedia to have flexible height per documentation
The documentation for CardMedia specifies that if the `aspectRatio`
property is not set, the media will have a flexible height.  The
existing implementation did not provide this, and would instead always
render with a height of `0`, effectively hiding the element.

This change should bring the component into line with the documentation.