Change autocomplete filter behaviour when multiple=false
Adds a showAllSuggestions state, which is set to true initially.
It is set to false when the input is changed, and is used to control if
the shown suggestions should be filtered based when the multiple prop is false.
Once a value is set, the state will be reset to true.