add 'onActiveChange' callback to DatePicker
this makes external management of the 'active' property much simpler

for example - the onFocus event was cancelling the onClick event for the text box

the onFocus handler did not trigger the 'onClick' property and so there was no way to properly manage the 'active' property externally.

this commit adds a callback for whenever the 'active' state changes internally so the external state (e.g. in redux can be updated also)

the result should be:

 * whenever the internal state changes - the external callback triggers updating external state to the internal value
 * if the external state changes - the new value should be passed in as the 'active' property and this.componentWillReceiveProps will call this.setState

these changes should be backwards compatable with existing code as they don't change existing callback handlers or the internal triggers just adds an extra callback