Format code using prettier
This was mostly automatic. The only catch was a couple places where we
use comments to disable eslint rules and those comments got wrapped to a
new line. In those few cases, I move the comment to the line above and
changed it from `eslint-disable-line` to `eslint-disable-next-line`.

This commit can be reproduced with:

  npm run lint:js -- --fix

Then manually fix the 5 or 6 places as described above. Running this
command again now should fix all the errors

  npm run lint:js -- --fix

Use the command from .lintstagedrc to format the TypeScript files

  prettier --write --print-width 100 --trailing-comma es5 --single-quote '**/*.ts'