Fix dropdown entry selection issue
Currently a user have to click twice on an entry within the dropdown
list to close it. This happens because the li click event will call
'handleSelect' and the default behavior calls 'handleDocumentClick'.
This is wrong when the user clicks on an entry (and not outside of
the dropdown) so now the 'handleSelect' prevents the default behaviour.