Allow tabs to be navigated with up and down keys
A guideline is to code the page such that the up and left
arrows move left in the tablist and the right and down arrows
move right to allow for user error. Screen reader users will
encounter the page linearly and will not know if the page tabs
are horizontal or vertical, only that appear one after the other.