Update README.md with proposed 3.0 version
Due to recent  breaking changes in the build requirements (i.e. ability to process color-mod) from versions currently published on npm, I propose releasing the next version as 3.0 (without the beta designation as it's quite stable) after merging our PR #1920.
This change adds a link to a new 2.0 to 3.0 migration guide I wrote. (I also reverted the changes to my 1.3 to 2.0 guide so it can be used with the currently published version on npm).
This change also specifies needing an older version of postcss-preset-env. More details are in the new 3.0 migration guide.